Fix issue that seek offset overflows int32

+++ storage/yottadisk.go
@@ -70,8 +70,7 @@ func (disk *YottaDisk) ReadData(dataIndex ydcommon.IndexTableValue) ([]byte, err
 
 	reader, err := disk.store.Reader()
 	dataBlock := make([]byte, disk.meta.DataBlockSize, disk.meta.DataBlockSize)
-	reader.Seek((int64)(disk.meta.DataOffset+
-		disk.meta.DataBlockSize*(uint32)(dataIndex)), io.SeekStart)
+	reader.Seek(int64(disk.meta.DataOffset)+int64(disk.meta.DataBlockSize)*int64(dataIndex), io.SeekStart)
 	_, err = reader.Read(dataBlock)
 	if err != nil {
 		return nil, err
@@ -93,8 +92,7 @@ func (disk *YottaDisk) WriteData(dataOffsetIndex ydcommon.IndexTableValue, data 
 	ydcommon.YottaAssert(len(data) <= (int)(disk.meta.DataBlockSize))
 	dataBlock := make([]byte, disk.meta.DataBlockSize, disk.meta.DataBlockSize)
 	copy(dataBlock, data)
-	writer.Seek((int64)(disk.meta.DataOffset+
-		disk.meta.DataBlockSize*(uint32)(dataOffsetIndex)), io.SeekStart)
+	writer.Seek(int64(disk.meta.DataOffset)+int64(disk.meta.DataBlockSize)*int64(dataOffsetIndex), io.SeekStart)
 	_, err = writer.Write(dataBlock)
 	if err != nil {
 		return err
